+++ tensor/tensor_test/forward_test/accessors_test.go
@@ -1,6 +1,10 @@
+// Accessors are the second group of functions under test in this package.
+// They are verified using only the primary functions (Full, Of, At, Equals)
+// established as a baseline in primary_ops_test.go.
 package forward_test
 
 import (
+	"fmt"
 	"slices"
 	"testing"
 
@@ -589,6 +593,452 @@ func TestSlice(t *testing.T) {
 			if err == nil {
 				t.Fatalf("expected error because of to index (3) being out of range [0,2) at dimension (0)")
 			} else if err.Error() != "Slice input index validation failed: expected index to fall in range [0,2] at dimension (0): got [1,3)" {
+				t.Fatal("unexpected error message returned")
+			}
+		})
+	})
+}
+
+func TestPatch(t *testing.T) {
+	tensor.RunTestLogicOnDevices(func(dev tensor.Device) {
+
+		// ============================== main paths ==============================
+
+		t.Run("Full(nil, 0.) scalar and Full(nil, 1.) scalar / Patch(nil) / replaces scalar value", func(t *testing.T) {
+			t1, err := tensor.Full(nil, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full(nil, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch(nil, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of(1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([2], 0.) and Full([1], 1.) / Patch([{1,2}]) / patches last element", func(t *testing.T) {
+			t1, err := tensor.Full([]int{2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{1}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 1, To: 2}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([]float64{0., 1.}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([3,3], 0.) and Full([2,2], 1.) / Patch([{0,2},{0,2}]) / patches top-left 2x2 block", func(t *testing.T) {
+			t1, err := tensor.Full([]int{3, 3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{2, 2}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 0, To: 2}, {From: 0, To: 2}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][]float64{
+				{1., 1., 0.},
+				{1., 1., 0.},
+				{0., 0., 0.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([3,3], 0.) and Full([2,2], 1.) / Patch([{0,2},{1,3}]) / patches top-right 2x2 block", func(t *testing.T) {
+			t1, err := tensor.Full([]int{3, 3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{2, 2}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 0, To: 2}, {From: 1, To: 3}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][]float64{
+				{0., 1., 1.},
+				{0., 1., 1.},
+				{0., 0., 0.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([3,3], 0.) and Full([2,2], 1.) / Patch([{1,3},{0,2}]) / patches bottom-left 2x2 block", func(t *testing.T) {
+			t1, err := tensor.Full([]int{3, 3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{2, 2}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 1, To: 3}, {From: 0, To: 2}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][]float64{
+				{0., 0., 0.},
+				{1., 1., 0.},
+				{1., 1., 0.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([3,3], 0.) and Full([2,2], 1.) / Patch([{1,3},{1,3}]) / patches bottom-right 2x2 block", func(t *testing.T) {
+			t1, err := tensor.Full([]int{3, 3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{2, 2}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 1, To: 3}, {From: 1, To: 3}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][]float64{
+				{0., 0., 0.},
+				{0., 1., 1.},
+				{0., 1., 1.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([4,3,2], 0.) and Full([3,2,1], 1.) / Patch(nil) / patches from origin with nil ranges", func(t *testing.T) {
+			t1, err := tensor.Full([]int{4, 3, 2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{3, 2, 1}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch(nil, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][][]float64{
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+				{
+					{0., 0.},
+					{0., 0.},
+					{0., 0.},
+				},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([4,3,2], 0.) and Full([3,2,1], 1.) / Patch([{1,4}]) / patches from dim-0 offset 1", func(t *testing.T) {
+			t1, err := tensor.Full([]int{4, 3, 2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{3, 2, 1}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 1, To: 4}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][][]float64{
+				{
+					{0., 0.},
+					{0., 0.},
+					{0., 0.},
+				},
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+				{
+					{1., 0.},
+					{1., 0.},
+					{0., 0.},
+				},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Full([4,3,2], 0.) and Full([3,2,1], 1.) / Patch([{1,4},{1,3},{1,2}]) / patches from all-dimension offset", func(t *testing.T) {
+			t1, err := tensor.Full([]int{4, 3, 2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{3, 2, 1}, 1., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch([]tensor.Range{{From: 1, To: 4}, {From: 1, To: 3}, {From: 1, To: 2}}, t2)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][][]float64{
+				{
+					{0., 0.},
+					{0., 0.},
+					{0., 0.},
+				},
+				{
+					{0., 0.},
+					{0., 1.},
+					{0., 1.},
+				},
+				{
+					{0., 0.},
+					{0., 1.},
+					{0., 1.},
+				},
+				{
+					{0., 0.},
+					{0., 1.},
+					{0., 1.},
+				},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		t.Run("Of([[1,2],[3,4]]) tensor / Patch(nil, self) / returns same tensor", func(t *testing.T) {
+			t1, err := tensor.Of([][]float64{
+				{1., 2.},
+				{3., 4.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			act, err := t1.Patch(nil, t1)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			exp, err := tensor.Of([][]float64{
+				{1., 2.},
+				{3., 4.},
+			}, &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if eq, err := act.Equals(exp); err != nil {
+				t.Fatal(err)
+			} else if !eq {
+				t.Fatalf("expected tensors to be equal")
+			}
+		})
+
+		// ============================== validations ==============================
+
+		t.Run("Full([2], 0.) / Patch([], nil) / returns error: source not on device", func(t *testing.T) {
+			t1, err := tensor.Full([]int{2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			_, err = t1.Patch([]tensor.Range{}, nil)
+			if err == nil {
+				t.Fatalf("expected error because of nil input tensor")
+			} else if err.Error() != fmt.Sprintf("Patch tensors' device validation failed: expected input tensor to be on %s", dev) {
+				t.Fatal("unexpected error message returned")
+			}
+		})
+
+		t.Run("Full([2], 0.) and Full(nil, 0.) scalar / Patch([]) / returns error: dimension count mismatch", func(t *testing.T) {
+			t1, err := tensor.Full([]int{2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full(nil, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			_, err = t1.Patch([]tensor.Range{}, t2)
+			if err == nil {
+				t.Fatalf("expected error because of incompatible number of dimensions")
+			} else if err.Error() != "Patch input index or tensors' dimension validation failed: expected number of dimensions to match among source and target tensors: (0) != (1)" {
+				t.Fatal("unexpected error message returned")
+			}
+		})
+
+		t.Run("Full([1,1], 0.) and Full([1,2], 0.) / Patch([]) / returns error: source exceeds target at dim 1", func(t *testing.T) {
+			t1, err := tensor.Full([]int{1, 1}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{1, 2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			_, err = t1.Patch([]tensor.Range{}, t2)
+			if err == nil {
+				t.Fatalf("expected error because of exceeding patch size at dimension (1)")
+			} else if err.Error() != "Patch input index or tensors' dimension validation failed: expected source tensor size not to exceed that of target tensor at dimension (1): (2) > (1)" {
+				t.Fatal("unexpected error message returned")
+			}
+		})
+
+		t.Run("Full([3], 0.) and Full([2], 0.) / Patch([{2,4}]) / returns error: index out of target range", func(t *testing.T) {
+			t1, err := tensor.Full([]int{3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			_, err = t1.Patch([]tensor.Range{{From: 2, To: 4}}, t2)
+			if err == nil {
+				t.Fatalf("expected error because of incompatible index with target tensor")
+			} else if err.Error() != "Patch input index or tensors' dimension validation failed: index incompatible with target tensor: expected index to fall in range [0,3] at dimension (0): got [2,4)" {
+				t.Fatal("unexpected error message returned")
+			}
+		})
+
+		t.Run("Full([1,3], 0.) and Full([1,2], 0.) / Patch([{},{2,3}]) / returns error: index does not cover source at dim 1", func(t *testing.T) {
+			t1, err := tensor.Full([]int{1, 3}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+			t2, err := tensor.Full([]int{1, 2}, 0., &tensor.Config{Device: dev})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			_, err = t1.Patch([]tensor.Range{{}, {From: 2, To: 3}}, t2)
+			if err == nil {
+				t.Fatalf("expected error because of index not covering source tensor at dimension (1)")
+			} else if err.Error() != "Patch input index or tensors' dimension validation failed: expected index to exactly cover source tensor at dimension (1): #[2,3) != (2)" {
 				t.Fatal("unexpected error message returned")
 			}
 		})
